fix(nodes): migrate exa search to @tool_function pattern and address review feedback
- Delete exa_driver.py and move search logic into IInstance.py using
  @tool_function decorator, matching the tool_firecrawl pattern
- Add output_schema and summary to the @tool_function decorator
- Fix bare Exception in IGlobal.py: use rocketlib.error() + ValueError
- Store config values directly on IGlobal instead of the driver object
- Add test config with default profile to services.json
